refactor groupscreen
