+++ app/(tabs)/groups.tsx
@@ -1,2 +1,2 @@
-import GroupScreen from '../screens/GroupScreen';
+import GroupScreen from '../screens/group/GroupScreen';
 export default GroupScreen;
