+++ app/group-members.tsx
@@ -1,3 +1,3 @@
-import GroupMembersScreen from './screens/GroupMembersScreen';
+import GroupMembersScreen from './screens/group/components/GroupMembersScreen';
 
 export default GroupMembersScreen;
